split MessageHandlers into two files move MessageHandlers into helpers update import statements remove unused import
