+++ src/pages/chat/components/FormattedText.tsx
@@ -3,7 +3,7 @@ import {
 	HandleNonNumberedList,
 	HandleNumberedList,
 	HandleTableResponse,
-} from "@/components/messageHandler/MessageHandlers";
+} from "@/helpers/MessageHandlers";
 import React from "react";
 import styles from "../Chat.module.css";
 import { Item } from "../[name]";
